+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## next
+
 ## 1.11.0
 
 - Add syntax highlight for `meson.options` files.